LayoutTests/fast/dom/HTMLTemplateElement/cycles-in-shadow-expected.txt

Test that cycle detection traverses over both templates and shadow roots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowDiv.appendChild(outerDiv) threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS successfullyParsed is true

TEST COMPLETE
